options: add support of using any file as linker script
Modify logic of LINKER_SCRIPT_FILE=xxx, the file provided in this option
now will be sent directly to linker.
Add LINKER_USE_TEMPLATE=0|1 toggle, when set to 1, LINKER_SCRIPT_FILE
will be treated as a template and be sent to compiler. It then generates .ldf
file as linker script.

Signed-off-by: Yuguo Zou <yuguo.zou@synopsys.com>